Initial version of executor Chance envp logical to char ** Create others execs
